test(query-builder): move take regression tests into query-builder/take/
