+++ test/functional/query-builder/take/mysql-skip-sql/entity/Animal.ts
@@ -1,0 +1,21 @@
+import {
+    Column,
+    Entity,
+    PrimaryGeneratedColumn,
+} from "../../../../../../src/index"
+import { Category } from "./Category"
+import { JoinTable } from "../../../../../../src/decorator/relations/JoinTable"
+import { ManyToMany } from "../../../../../../src/decorator/relations/ManyToMany"
+
+@Entity()
+export class Animal {
+    @PrimaryGeneratedColumn()
+    id: number
+
+    @Column()
+    name: string
+
+    @ManyToMany(() => Category)
+    @JoinTable()
+    categories: Category[]
+}
