+++ test/functional/query-builder/take/mysql-skip-sql/entity/Category.ts
@@ -1,0 +1,7 @@
+import { Entity, PrimaryGeneratedColumn } from "../../../../../../src/index"
+
+@Entity()
+export class Category {
+    @PrimaryGeneratedColumn()
+    id: number
+}
