+++ test/functional/query-builder/take/string-primary-key/entity/Post.ts
@@ -1,0 +1,12 @@
+import { Entity } from "../../../../../../src/decorator/entity/Entity"
+import { Column } from "../../../../../../src/decorator/columns/Column"
+import { PrimaryColumn } from "../../../../../../src/decorator/columns/PrimaryColumn"
+
+@Entity()
+export class Post {
+    @PrimaryColumn()
+    id: string
+
+    @Column()
+    title: string
+}
