+++ test/functional/query-builder/take/with-ids-parameter/entity/User.ts
@@ -1,0 +1,12 @@
+import { Entity } from "../../../../../../src/decorator/entity/Entity"
+import { PrimaryColumn, ManyToOne } from "../../../../../../src"
+import { Role } from "./Role"
+
+@Entity()
+export class User {
+    @PrimaryColumn()
+    id: number
+
+    @ManyToOne((_) => Role, (role) => role.users)
+    role: Role
+}
