+++ test/functional/query-builder/take/with-ids-parameter/with-ids-parameter.test.ts
@@ -1,0 +1,57 @@
+import "reflect-metadata"
+import { expect } from "chai"
+import type { DataSource } from "../../../../../src"
+import { User } from "./entity/User"
+import { Role } from "./entity/Role"
+import {
+    createTestingConnections,
+    reloadTestingDatabases,
+    closeTestingConnections,
+} from "../../../../utils/test-utils"
+
+describe("query-builder > take > with ids parameter", () => {
+    let dataSources: DataSource[]
+
+    before(async () => {
+        dataSources = await createTestingConnections({
+            disabledDrivers: ["spanner"],
+            entities: [User, Role],
+            schemaCreate: true,
+            dropSchema: true,
+        })
+    })
+    beforeEach(() => reloadTestingDatabases(dataSources))
+    after(() => closeTestingConnections(dataSources))
+
+    it("should allow the 'ids' parameter without messing up the query when using .take", () =>
+        Promise.all(
+            dataSources.map(async (connection) => {
+                const roleRepository = connection.getRepository(Role)
+                const userRepository = connection.getRepository(User)
+
+                const userWithId = (id: number) => ({ id })
+
+                await roleRepository.save([
+                    { id: "a", users: [1, 2, 3].map(userWithId) },
+                    { id: "b", users: [9, 10, 11, 12, 13].map(userWithId) },
+                    { id: "c", users: [14, 15, 16, 17].map(userWithId) },
+                ])
+
+                const results = await userRepository
+                    .createQueryBuilder("user")
+                    .leftJoinAndSelect("user.role", "role")
+                    .where("role.id IN (:...ids)", { ids: ["a", "c"] })
+                    .take(5)
+                    .orderBy("user.id")
+                    .getMany()
+
+                expect(results).to.be.deep.equal([
+                    { id: 1, role: { id: "a" } },
+                    { id: 2, role: { id: "a" } },
+                    { id: 3, role: { id: "a" } },
+                    { id: 14, role: { id: "c" } },
+                    { id: 15, role: { id: "c" } },
+                ])
+            }),
+        ))
+})
