--- test/github-issues/1099/entity/Category.ts
@@ -1,7 +1,0 @@
-import { Entity, PrimaryGeneratedColumn } from "../../../../src/index"
-
-@Entity()
-export class Category {
-    @PrimaryGeneratedColumn()
-    id: number
-}
