--- test/github-issues/493/entity/Post.ts
@@ -1,12 +1,0 @@
-import { Entity } from "../../../../src/decorator/entity/Entity"
-import { Column } from "../../../../src/decorator/columns/Column"
-import { PrimaryColumn } from "../../../../src/decorator/columns/PrimaryColumn"
-
-@Entity()
-export class Post {
-    @PrimaryColumn()
-    id: string
-
-    @Column()
-    title: string
-}
