--- test/github-issues/5174/entity/User.ts
@@ -1,12 +1,0 @@
-import { Entity } from "../../../../src/decorator/entity/Entity"
-import { PrimaryColumn, ManyToOne } from "../../../../src"
-import { Role } from "./Role"
-
-@Entity()
-export class User {
-    @PrimaryColumn()
-    id: number
-
-    @ManyToOne((_) => Role, (role) => role.users)
-    role: Role
-}
