--- test/other-issues/limit-with-order-by/entity/Category.ts
@@ -1,12 +1,0 @@
-import { Entity } from "../../../../src/decorator/entity/Entity"
-import { PrimaryGeneratedColumn } from "../../../../src/decorator/columns/PrimaryGeneratedColumn"
-import { Column } from "../../../../src/decorator/columns/Column"
-
-@Entity()
-export class Category {
-    @PrimaryGeneratedColumn()
-    id: number
-
-    @Column()
-    name: string
-}
